Added the ConfigureDefaults API, as an extension method on the IHostBuilder interface. Fix #36003
